[ADDED]: first working version

+++ app/src/main/java/com/example/alias/presentation/viewmodels/ConfigViewModel.kt
@@ -9,7 +9,7 @@ class ConfigViewModel : ViewModel() {
         Game.winScore = score
     }
 
-    fun setPlaytime(time: Int) {
+    fun setPlaytime(time: Long) {
         Game.playtime = time
     }
 }
